medical-chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pydantic>=2
sentence-transformers==2.2.2
chromadb==1.5.5
streamlit>=1.28.0
pandas>=2.0.0